+++ backend/accounts/admin.py
@@ -1,2 +1,10 @@
+from django.contrib import admin
 
-# Register your models here.
+from .models import User
+
+
+@admin.register(User)
+class UserAdmin(admin.ModelAdmin):
+    list_display = ["email", "first_name", "is_staff", "is_active", "date_joined"]
+    search_fields = ["email", "first_name"]
+    exclude = ["password"]